backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-core==2.20.1
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.42.0
google-generativeai==0.4.0
anthropic==0.30.0
aiohttp==3.10.5
#rouge-score==0.1.2
#nltk==3.9
sentence-transformers==3.0.1
scikit-learn==1.5.1
huggingface-hub==0.24.5

